NOTE: May not be backwards compatible
It was a long time ago, I don't really know why I did -x
the way I did. It seems to be unecessary to ever delete
the record before update, and more likely to simply want
to delete a record[s] period.

I've changed the option to make the latter easier at the
expense of the former. The former behavior still happens
*IF* a value is explicitely given and given AFTER all -x
options.

I'd be willing to revert the -x behavior and make a -X
for the new behavior if this greatly offends anyone.

More likely is that nobody else uses this.